+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/MemoryBatchBaseDao.java
@@ -7,15 +7,13 @@ import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
 import org.springframework.lang.NonNull;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
 /**
  * 通过内存实现的基础数据访问层。
+ *
  * <p>该类只提供最基本的方法实现，没有添加同步锁或其它的安全性一致性保证，请通过代理的方式在代理类中添加。</p>
  * <p>该类可以通过 MapResourceBridge 对内存映射进行外部资源桥接。
  * 在数据访问层启动前读取资源数据，在结束之后保存数据，即可实现内存数据的持久化。</p>
@@ -27,6 +25,10 @@ import java.util.stream.Collectors;
 public class MemoryBatchBaseDao<K extends Key, E extends Entity<K>> implements BatchBaseDao<K, E> {
 
     private Map<K, E> memory;
+
+    public MemoryBatchBaseDao() {
+        this(new HashMap<>());
+    }
 
     public MemoryBatchBaseDao(@NonNull Map<K, E> memory) {
         this.memory = memory;
@@ -129,6 +131,34 @@ public class MemoryBatchBaseDao<K extends Key, E extends Entity<K>> implements B
             elements.add(memory.get(key));
         }
         return elements;
+    }
+
+    /**
+     * 填充数据。
+     *
+     * @param mrb 映射资源桥。
+     * @throws DaoException 数据访问层异常。
+     */
+    public void fillData(@NonNull MapResourceBridge<K, E> mrb) throws DaoException {
+        try {
+            mrb.fillMap(this.memory);
+        } catch (Exception e) {
+            throw new DaoException(e);
+        }
+    }
+
+    /**
+     * 保存数据。
+     *
+     * @param mrb 映射资源桥。
+     * @throws DaoException 数据访问层异常。
+     */
+    public void saveData(@NonNull MapResourceBridge<K, E> mrb) throws DaoException {
+        try {
+            mrb.saveMap(this.memory);
+        } catch (Exception e) {
+            throw new DaoException(e);
+        }
     }
 
     public Map<K, E> getMemory() {